+++ agent-control/tests/on_host/scenarios.rs
@@ -13,4 +13,5 @@ mod multiple_remote_config;
 mod no_orphan_processes;
 mod oci_auth;
 mod opamp;
+mod remote_agent_removal;
 mod restarting_processes;
